Checking  some alternatives

+++ 0x14-bit_manipulation/main.h
@@ -2,7 +2,7 @@
 #define MAIN_H
 
 unsigned int binary_to_uint(const char *b);
-int _pow(int base, int exponent);
+double _pow(double base, unsigned int exponent);
 int _strlen(const char *s);
 
 #endif
